+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [10.1.0] - 2024-02-02
+
 ### Added
 - Support extending schema using partial GraphQL schema files.